+++ src/compute-engine/compilation/interval-javascript-target.ts
@@ -185,6 +185,9 @@ const INTERVAL_JAVASCRIPT_FUNCTIONS: CompiledFunctions<Expression> = {
   Arcsch: (args, compile) => `_IA.acsch(${compile(args[0])})`,
   Arsech: (args, compile) => `_IA.asech(${compile(args[0])})`,
 
+  // Special functions
+  Gamma: (args, compile) => `_IA.gamma(${compile(args[0])})`,
+
   // Elementary
   Fract: (args, compile) => `_IA.fract(${compile(args[0])})`,
   Truncate: (args, compile) => `_IA.trunc(${compile(args[0])})`,